LRCI-3756 Move test class results methods to its own util class
