+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/AntUtil.java
@@ -253,8 +253,8 @@ public class AntUtil {
 			if (runningModulesTests) {
 				String projectName = parameters.get("test.task");
 
-				JenkinsResultsParserUtil.moveTestResultFiles(
-					JenkinsResultsParserUtil.getProjectTestResultDir(
+				TestClassResultUtil.moveTestResultFiles(
+					TestClassResultUtil.getProjectTestResultDir(
 						projectName, baseDir),
 					baseDir);
 			}
