+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/JenkinsResultsParserUtil.java
@@ -2705,22 +2705,6 @@ public class JenkinsResultsParserUtil {
 		}
 	}
 
-	public static File getProjectTestResultDir(
-		String projectName, File baseDir) {
-
-		Matcher matcher = _projectNamePattern.matcher(projectName);
-
-		if (matcher.find()) {
-			String testProjectName = matcher.group(1);
-
-			testProjectName = testProjectName.replaceAll(":", "/");
-
-			return new File(baseDir, "modules/" + testProjectName);
-		}
-
-		return null;
-	}
-
 	public static Properties getProperties(File... propertiesFiles) {
 		Properties properties = new Properties();
 
@@ -3253,17 +3237,6 @@ public class JenkinsResultsParserUtil {
 		return subdirectories;
 	}
 
-	public static String getTestResultFileName(File file) {
-		Matcher matcher = _testResultFileNamePattern.matcher(
-			file.getAbsolutePath());
-
-		if (matcher.find()) {
-			return matcher.group(0);
-		}
-
-		return null;
-	}
-
 	public static File getUserHomeDir() {
 		return _userHomeDir;
 	}
@@ -3705,30 +3678,6 @@ public class JenkinsResultsParserUtil {
 
 		if (!delete(sourceFile)) {
 			throw new IOException("Unable to delete " + sourceFile);
-		}
-	}
-
-	public static void moveTestResultFiles(
-			File testProjectResultsDir, File baseDir)
-		throws IOException {
-
-		String[] globList = {"**/TEST-**.xml"};
-
-		List<File> fileList = getIncludedFiles(
-			testProjectResultsDir, null, globList);
-
-		File testResultsDir = new File(baseDir, "modules/test-results");
-
-		for (File resultFile : fileList) {
-			String testResultFileName = getTestResultFileName(resultFile);
-
-			File destinationFile = new File(testResultsDir, testResultFileName);
-
-			move(resultFile, destinationFile);
-
-			if (!resultFile.exists()) {
-				System.out.println("Deleted result file:" + resultFile);
-			}
 		}
 	}
 
@@ -6266,8 +6215,6 @@ public class JenkinsResultsParserUtil {
 		"\\$\\{([^\\}]+)\\}");
 	private static final Pattern _poshiFileNamePattern = Pattern.compile(
 		".*\\.(function|macro|path|prose|testcase)");
-	private static final Pattern _projectNamePattern = Pattern.compile(
-		"(:.*)(:test|testIntegration)");
 	private static final Set<String> _redactTokens = new HashSet<>();
 	private static final Pattern _remoteURLAuthorityPattern1 = Pattern.compile(
 		"https://(test).liferay.com/([0-9]+)/");
@@ -6289,8 +6236,6 @@ public class JenkinsResultsParserUtil {
 		}
 	};
 
-	private static final Pattern _testResultFileNamePattern = Pattern.compile(
-		"TEST-.*.xml");
 	private static final Set<String> _timeStamps = new HashSet<>();
 	private static final Pattern _urlQueryStringPattern = Pattern.compile(
 		"\\&??(\\w++)=([^\\&]*)");
